- All unit tests pass except test_execute_action_with_handler_error
