+++ app/actions/tests/test_handlers.py
@@ -1,4 +1,5 @@
 import pytest
+
 from app.actions.rmwhub import RmwSets
 
 
@@ -39,6 +40,10 @@ async def test_handler_action_pull_observations(
     mocker.patch(
         "app.actions.rmwhub.RmwHubAdapter.process_rmw_download",
         return_value=mock_rmw_observations,
+    )
+    mocker.patch(
+        "app.actions.rmwhub.RmwHubAdapter.process_rmw_upload",
+        return_value=[],
     )
     mocker.patch(
         "app.actions.rmwhub.RmwHubAdapter.push_status_updates",
